+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Sastre 1.20 [xx, 2022]
 
 #### Enhancements:
 - [#42] Option --archive added to backup and restore tasks. Allowing backups to be saved as a zip file, and restore to read from a zip file.
+- [#164] Implemented adaptive mechanism in the REST API client, to backoff and retry when rate-limit signal (429) is received from vManage. 
 
 #### Fixes:
 - [#165] CustomApp policies failed to be restored due to dependency resolution issues, this has been fixed.